app/templates/server-tpl/requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-OAuthlib==0.8.0
Flask-RESTful==0.3.1
oauthlib==0.7.2
pydash==2.4.2
Werkzeug==0.10.1
factory_boy==2.4.1
factoryboy-gaendb==0.1.5
Unidecode==0.4.17